Implement matrix ranges etc.
Matrix ranges define a "view" on a matrix, they have their own dimensions and
can be operate on as if they were matrices of that size, but when modified,
they actually change the underlying matrix internal representation (and thus
the underlying matrix itself).

In this draft implementation only "happy" scenarios are covered.

Details:
    - Implement range(), col(), row() functions which return "ranges". col()
      and row() being special cases of range().
    - Implement Lower and Upper classes as "views" on a Matrix.
    - Re-implement LU factorization using ranges and Lower/Upper classes.
    - Add and fix tests.